Adress Working

+++ src/main/java/com/example/demo/models/Product.java
@@ -13,7 +13,7 @@ import jakarta.persistence.Table;
 public class Product {
     @Id
     @GeneratedValue( strategy = GenerationType.SEQUENCE)
-    private Integer id;
+    private Long id;
     private String name;
     private String description;
     private Double price;
@@ -22,7 +22,7 @@ public class Product {
     public Product() {
     }
 
-    public Product(Integer id, String name, String description, Double price) {
+    public Product(Long id, String name, String description, Double price) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -30,11 +30,11 @@ public class Product {
     }
 
 
-    public Integer getId() {
+    public Long getId() {
         return this.id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
